moved database to new file
