+++ README.md
@@ -81,14 +81,15 @@ python illumina_uploader.py --config config.ini --sequencer miseq --dry-run
      - ~~Generate update.txt and ignore.txt files~~
      - ~~Generate .json file for each run directory~~
      - ~~Generate COPY_COMPLETE file for each run directory~~
-- v0.2   - Email functionality using mail folder
+- v0.2
+     - Email functionality through plover using mail folder
+     - JSON status file: file status, checksum, num_files, timestamp
+     - run instrument error files send via email
 - v0.3   - Web UI using flask/django
-- v0.4   - Email run instrument errors (Can be done on cloud?)
-- v0.5   - Upload to IRIDA for full analysis. Get IRIDA project number from samplesheet. Check sequdas code.
-- v0.6   - QC metrics (separate MultiQC and Kraken summery screens)
-- v0.7   - Advanced Data integrity check
-- v0.8   - Installer and one script run
-- v0.9   - Progress bar in UI using API
+- v0.4   - Upload to IRIDA for full analysis. Get IRIDA project number from samplesheet. Check sequdas code.
+- v0.5   - Advanced Data integrity check
+- v0.6   - Installer and one script run
+- v0.7   - Progress bar in UI using API
 - v1.0   - First Release Freeze, no new features
 
 ## Troubleshooting
